requirements.txt
fastapi
uvicorn
python-dotenv
razorpay
pyserial
requests
pydantic
email-validator
python-multipart
